+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.13] - 2026-07-23
+
+### Added
+
+- Command group `playlist` with the `list` and `play` commands
+- Subsection for the `playlist list` subcommand of the `output` section
+  of the configuration file
+
+
 ## [0.0.12] - 2026-07-23
 
 ### Added